pyproject.toml
[tool.poetry]
name = "blacklist"
version = "1.9.0"
description = ""
authors = ["Dmitriev Andrey <justkeepwalking@yandex.ru>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
pydantic-settings = "^2.1.0"
orjson = "^3.9.10"
redis = "^5.0.1"
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
celery = "^5.3.6"
tomli = "^2.0.1"

[tool.poetry.group.dev.dependencies]
coverage = "^7.4.1"
flake8 = "^7.0.0"
mypy = "^1.8.0"
isort = "^5.13.2"
black = "^24.1.1"
pre-commit = "^3.6.0"
flake8-bugbear = "^24.1.17"
pep8-naming = "^0.13.3"
flake8-builtins = "^2.2.0"
flake8-quotes = "^3.3.2"


[tool.poetry.group.jupyter.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 120
force_single_line = true

[tool.mypy]
python_version = 3.11
plugins = ["pydantic.mypy"]
check_untyped_defs = true
exclude = [
    ".*/__to_delete/.*\\.py",
]

[tool.black]
line-length = 120
target-version = ["py311"]
skip-string-normalization = true
